refactor: move 2 TM utils out
