+++ packages/maskbook/src/plugins/External/components/CompositionEntry.tsx
@@ -1,6 +1,6 @@
-import type { Plugin } from '@masknet/plugin-infra/src'
-import { usePortalShadowRoot } from '@masknet/shared'
-import { MaskDialog } from '@masknet/theme'
+import type { Plugin } from '@masknet/mask-plugin-infra'
+import { usePortalShadowRoot } from '@masknet/maskbook-shared'
+import { MaskDialog } from '@masknet/maskbook-theme'
 import { DialogContent } from '@material-ui/core'
 import { PluginLoader } from './PluginLoader'
